requirements.txt
openai
elevenlabs
pydub
python-dotenv
argparse
requests
TTS
fastapi
uvicorn
pydantic